pyproject.toml
[tool.poetry]
name = "uzi4d"
version = "0.1.0"
description = ""
authors = [
    "AsuraBot <homecat13@gmail.com>",
    "GurgenXD <gurgen.abramyan.1995@gmail.com>"
]
readme = "README.md"
packages = [
    {include = "uzi4d"}
]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
pillow = "^9.4.0"
requests = "^2.28.2"
django-admin-extra-buttons = "^1.5.6"
pydantic = "^1.10.5"
psycopg2 = "^2.9.5"
django-mptt = "^0.14.0"
django-mptt-admin = "^2.4.1"

[tool.poetry.group.dev.dependencies]
black = "23.1.0"
pytest = "7.2.1"
pytest-cov = "4.0.0"
pytest-deadfixtures = "2.2.1"
ruff = "0.0.246"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]
skip-magic-trailing-comma = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["ALL"]
ignore = [
    "RUF001", "RUF002", "RUF003", "D104", "D100", "D107", "COM812", "ANN101", "D106",
    "A002", "A003", "FBT003", "D105", "TCH003", "TCH002", "D203", "D213", "UP017",
    "FBT001", "ANN401", "FBT002", "ANN102", "G004", "PGH003", "TRY003", "EM101", "ANN001",
    "ANN201", "ANN206", "ANN204"
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"app/migrations/**/*.py" = ["D101"]
"app/utils/query.py" = ["ALL"]

[tool.ruff.isort]
lines-after-imports = 2

[tool.ruff.pylint]
max-args = 10
